Implementing omrs-link.
