+++ .prettierignore
@@ -1,0 +1,1 @@
+src/buttons-and-links/buttons-and-links.stories.html
