--- src/button/button.stories.html
@@ -1,137 +1,0 @@
-<article class="stories-button">
-  <h2>Buttons</h2>
-  <section>
-    <p>
-      <span class="omrs-underline omrs-italic omrs-bold"
-        >You should use buttons more often than you're used to.</span
-      >
-    </p>
-    <p>
-      If it has a click listener, it should probably be a button.
-    </p>
-    <div class="stories-example lang-html">
-      <!-- Don't put click listeners on divs!! Ever!! -->
-      <!-- They don't participate in keyboard navigation nor are usable by screenreaders! -->
-      <!-- 😢😢  😭😭  🤢🤢  🤮🤮  ❌❌ -->
-      <div onclick="">Please don't do this</div>
-
-      <!-- Don't put href on buttons!! Or divs!! Ever!! -->
-      <!-- An href means it's a link. Links should be done with <a> -->
-      <!-- 🤢🤢  ⛔⛔  😓😓  💀💀 -->
-      <button href="/some-link">This is wrong</button>
-
-      <!-- Things that have a click listener but no href should be buttons -->
-      <!-- If it doesn't look like a normal button, add "omrs-unstyled" to it. -->
-      <!-- Buttons participate perfectly in keyboard navigation and are usable by screen readers! -->
-      <!-- Also, keyboard navigation and screen readers will love you. -->
-      <!-- 👏👏  👍👍  🆗🆗  👌👌 -->
-      <button onclick="" class="omrs-unstyled">Do this!</button>
-
-      <!-- Things that have a click listener but no href should be buttons -->
-      <!-- Using <a> enables Ctrl/Cmd + Click to open in new tab. -->
-      <!-- <a> participates perfectly in keyboard navigation and is totally usable by screen readers! -->
-      <!-- 🙌🙌 🔥🔥  🥰🥰  ❤️❤️ -->
-      <a href="/some-link">Do this!</a>
-
-      <!-- Using the role attribute is okay, but this is easier to mess up -->
-      <!-- If you choose to use role="button", make sure to set tabindex -->
-      <!-- 😐😐 -->
-      <div role="button" tabindex="0">Okay, but use omrs-unstyled instead</div>
-    </div>
-  </section>
-  <h3>Style types</h3>
-  <div>
-    <h4>Unstyled Buttons</h4>
-    <section class="button-examples">
-      <button class="omrs-unstyled">
-        Unstyled button
-      </button>
-    </section>
-  </div>
-  <div>
-    <h4>Action Buttons</h4>
-    <section class="button-examples">
-      <button class="omrs-btn omrs-filled-action">
-        Filled button
-      </button>
-      <button class="omrs-btn omrs-outlined-action">
-        Outlined button
-      </button>
-      <button class="omrs-btn omrs-text-action">
-        Text button
-      </button>
-      <button class="omrs-btn omrs-filled-action omrs-rounded">
-        Rounded button
-      </button>
-    </section>
-  </div>
-  <div>
-    <h4>Neutral Buttons</h4>
-    <section class="button-examples">
-      <button class="omrs-btn omrs-filled-neutral">
-        Filled button
-      </button>
-      <button class="omrs-btn omrs-outlined-neutral">
-        Outlined button
-      </button>
-      <button class="omrs-btn omrs-text-neutral">
-        Text button
-      </button>
-      <button class="omrs-btn omrs-filled-neutral omrs-rounded">
-        Rounded button
-      </button>
-    </section>
-  </div>
-  <div>
-    <h4>Destructive Buttons</h4>
-    <section class="button-examples">
-      <button class="omrs-btn omrs-outlined-destructive">
-        Outlined button
-      </button>
-      <button class="omrs-btn omrs-text-destructive">
-        Text button
-      </button>
-      <button class="omrs-btn omrs-text-destructive omrs-rounded">
-        Text Rounded button
-      </button>
-    </section>
-  </div>
-  <div>
-    <h4>Disabled Buttons</h4>
-    <section class="button-examples">
-      <button class="omrs-btn omrs-filled-disabled" disabled>
-        Filled button
-      </button>
-      <button class="omrs-btn omrs-outlined-disabled" disabled>
-        Outlined button
-      </button>
-      <button class="omrs-btn omrs-text-disabled" disabled>
-        Text button
-      </button>
-    </section>
-  </div>
-  <section>
-    <h3>Code example:</h3>
-    <button
-      class="<%= determineStyle(buttonType, styleType, isRounded, isLarge) %>"
-    >
-      A button
-    </button>
-    <div class="stories-example lang-html">
-      <button
-        class="<%= determineStyle(buttonType, styleType, isRounded, isLarge) %>"
-      >
-        A button
-      </button>
-    </div>
-  </section>
-</article>
-<style>
-  .stories-button .button-examples {
-    display: flex;
-  }
-
-  .stories-button .button-examples > * {
-    margin: 0 0.5rem 1rem 0.5rem;
-  }
-</style>
