+++ src/styleguide-entry.js
@@ -4,7 +4,7 @@ import "./colors/colors.css";
 import "./typography/typography.css";
 import "./spacing/spacing.css";
 import "./cards/cards.css";
-import "./button/button.css";
+import "./buttons-and-links/buttons-and-links.css";
 import "./logo/logo.js";
 import "./icons/icons.js";
 import "./inputs/input.css";
